backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
Werkzeug==2.0.3
numpy==2.1.3
Pillow==10.2.0